move components_list.cjs to workspace common
