+++ packages/dialtone-vue3/.eslintrc.cjs
@@ -1,4 +1,4 @@
-const componentsList = require('./common/components_list.cjs');
+const componentsList = require('../../common/components_list.cjs');
 componentsList.push('btn', 'select', 'validation-message', 'label', 'description');
 const componentsNames = componentsList.map(name => name.replace('_', '-').replace('.vue', ''));
