--- packages/dialtone-vue3/common/components_list.cjs
@@ -1,50 +1,0 @@
-module.exports = [
-  'datepicker.vue',
-  'avatar.vue',
-  'badge.vue',
-  'banner.vue',
-  'breadcrumbs.vue',
-  'button.vue',
-  'button_group.vue',
-  'card.vue',
-  'checkbox.vue',
-  'checkbox_group.vue',
-  'chip.vue',
-  'codeblock.vue',
-  'collapsible.vue',
-  'combobox.vue',
-  'description_list.vue',
-  'dropdown.vue',
-  'emoji.vue',
-  'emoji_picker.vue',
-  'emoji_text_wrapper.vue',
-  'hovercard.vue',
-  'icon.vue',
-  'image_viewer.vue',
-  'input.vue',
-  'input_group.vue',
-  'item_layout.vue',
-  'keyboard_shortcut.vue',
-  'lazy_show.vue',
-  'link.vue',
-  'list_item.vue',
-  'modal.vue',
-  'notice.vue',
-  'pagination.vue',
-  'popover.vue',
-  'presence.vue',
-  'radio.vue',
-  'radio_group.vue',
-  'root_layout.vue',
-  'scroller.vue',
-  'select_menu.vue',
-  'skeleton.vue',
-  'stack.vue',
-  'tab.vue',
-  'tab_group.vue',
-  'tab_panel.vue',
-  'toast.vue',
-  'toggle.vue',
-  'tooltip.vue',
-  'validation_messages.vue',
-];
